responsive customize page
